Added ROS subset, where I look at linear relationship between early and late ROS subset by open VS closed canopy plots.
Next, I plan to use segmented function (as in grass load analysis) to identify breakpoints for early and late season fire effects (ROS, flame length, intensity). This will be limited analysis, merely identification of breakpoints.

Analyses will come in shape of the subset comparisons, because analysing across the entire range of canopy cover is statistically difficult and really all we are concerned with there is the breakpoint at which we go from flammable to not so flammable.